requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.1
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.4.0
cuopt-cu12==25.5.1
cuopt-mps-parser==25.5.1
cupy-cuda12x==13.4.1
cuvs-cu12==25.4.0
cvxpy==1.6.6
cycler==0.12.1
dask==2025.2.0
dask-cuda==25.4.0
datasets==4.0.0
dill==0.3.8
distributed==2025.2.0
distributed-ucxx-cu12==0.43.0
einops==0.8.1
fastrlock==0.8.3
filelock==3.18.0
flameprof==0.4
flash-attn==2.8.3
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
git-filter-repo==2.47.0
grpcio==1.73.1
grpcio-tools==1.73.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
libcudf-cu12==25.4.0
libcuopt-cu12==25.5.1
libcuvs-cu12==25.4.0
libkvikio-cu12==25.4.0
libraft-cu12==25.4.0
librmm-cu12==25.4.0
libucx-cu12==1.18.1
libucxx-cu12==0.43.0
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
numba==0.60.0
numba-cuda==0.4.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.12
osqp==1.0.4
packaging==25.0
pandas==2.2.2
partd==1.4.2
patsy==1.0.1
peft==0.16.0
pillow==11.3.0
pip @ file:///croot/pip_1746204010231/work
prometheus-client==0.22.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-grpc-prometheus==0.8.0
pyarrow==19.0.1
pycparser==2.22
pygments==2.19.2
pylibcudf-cu12==25.4.0
pylibraft-cu12==25.4.0
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
raft-dask-cu12==25.4.0
rapids-dask-dependency==25.4.0
rapids-logger==0.1.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rmm-cu12==25.4.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
scs==3.2.7.post2
sentencepiece==0.2.0
setuptools==78.1.1
six==1.17.0
sortedcontainers==2.4.0
statsmodels==0.14.5
sympy==1.14.0
tblib==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing-extensions==4.14.1
tzdata==2025.2
ucx-py-cu12==0.43.0
ucxx-cu12==0.43.0
urllib3==2.5.0
uv==0.7.20
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
